moved the component page of connect-and-join to a separate file
